refactor(search): split results-table component in smart and UI components
