+++ libs/feature/search/src/index.ts
@@ -21,4 +21,4 @@ export * from './lib/results-hits-number/results-hits.container.component'
 export * from './lib/results-layout/results-layout.component'
 export * from './lib/sort-by/sort-by.component'
 export * from './lib/state/container/search-state.container.directive'
-export * from './lib/results-table/results-table.component'
+export * from './lib/results-table/results-table-container.component'
