+++ libs/ui/search/src/index.ts
@@ -17,3 +17,4 @@ export * from './lib/record-preview-title/record-preview-title.component'
 export * from './lib/record-metric/record-metric.component'
 export * from './lib/results-list-item/results-list-item.component'
 export * from './lib/results-hits-number/results-hits-number.component'
+export * from './lib/results-table/results-table.component'
